requirements.txt
python-telegram-bot[callback-data]==21.6